refactor: migrate channel archival fn to conversation module
