+++ app/tests/modules/incident/test_incident_helper.py
@@ -143,82 +143,6 @@ def test_handle_incident_command_with_unknown_command():
     )
 
 
-@patch("modules.incident.incident_helper.log_to_sentinel")
-def test_archive_channel_action_ignore(mock_log_to_sentinel):
-    client = MagicMock()
-    body = {
-        "actions": [{"value": "ignore"}],
-        "channel": {"id": "channel_id", "name": "incident-2024-01-12-test"},
-        "message_ts": "message_ts",
-        "user": {"id": "user_id"},
-    }
-    ack = MagicMock()
-    respond = MagicMock()
-    incident_helper.archive_channel_action(client, body, ack, respond)
-    ack.assert_called_once()
-    client.chat_update.assert_called_once_with(
-        channel="channel_id",
-        text="<@user_id> has delayed scheduling and archiving this channel for 14 days.",
-        ts="message_ts",
-        attachments=[],
-    )
-    mock_log_to_sentinel.assert_called_once_with(
-        "incident_channel_archive_delayed", body
-    )
-
-
-@patch("modules.incident.incident_helper.close_incident")
-@patch("modules.incident.incident_helper.log_to_sentinel")
-def test_archive_channel_action_archive(
-    mock_log_to_sentinel,
-    mock_close_incident,
-):
-    client = MagicMock()
-    body = {
-        "actions": [{"value": "archive"}],
-        "channel": {"id": "channel_id", "name": "incident-2024-01-12-test"},
-        "message_ts": "message_ts",
-        "user": {"id": "user_id"},
-    }
-    channel_info = {
-        "channel_id": "channel_id",
-        "channel_name": "incident-2024-01-12-test",
-        "user_id": "user_id",
-    }
-    ack = MagicMock()
-    respond = MagicMock()
-    incident_helper.archive_channel_action(client, body, ack, respond)
-    assert ack.call_count == 1
-    mock_log_to_sentinel.assert_called_once_with("incident_channel_archived", body)
-    mock_close_incident.assert_called_once_with(client, channel_info, ack, respond)
-
-
-@patch("modules.incident.incident_helper.schedule_retro")
-@patch("modules.incident.incident_helper.log_to_sentinel")
-def test_archive_channel_action_schedule_incident(mock_log_to_sentinel, mock_schedule):
-    client = MagicMock()
-    body = {
-        "actions": [{"value": "schedule_retro"}],
-        "channel": {"id": "channel_id", "name": "incident-2024-01-12-test"},
-        "message_ts": "message_ts",
-        "user": {"id": "user_id"},
-        "trigger_id": "trigger_id",
-    }
-    channel_info = {
-        "channel_id": "channel_id",
-        "channel_name": "incident-2024-01-12-test",
-        "user_id": "user_id",
-        "trigger_id": "trigger_id",
-    }
-    ack = MagicMock()
-    incident_helper.archive_channel_action(client, body, ack, MagicMock())
-    assert ack.call_count == 1
-    mock_schedule.schedule_incident_retro.assert_called_once_with(
-        client, channel_info, ack
-    )
-    mock_log_to_sentinel.assert_called_once_with("incident_retro_scheduled", body)
-
-
 @patch("modules.incident.incident_helper.slack_channels.get_stale_channels")
 def test_stale_incidents(get_stale_channels_mock):
     client = MagicMock()
